Update feign samples (for 2.3.0 branch).
This removes some old files which are not generated anymore.

+++ samples/client/petstore/java/feign/src/test/java/io/swagger/client/api/FakeApiTest.java
@@ -1,10 +1,10 @@
 package io.swagger.client.api;
 
 import io.swagger.client.ApiClient;
-import io.swagger.client.model.Client;
-import org.threeten.bp.OffsetDateTime;
-import org.threeten.bp.LocalDate;
 import java.math.BigDecimal;
+import io.swagger.client.model.Client;
+import org.threeten.bp.LocalDate;
+import org.threeten.bp.OffsetDateTime;
 import org.junit.Before;
 import org.junit.Test;
 
@@ -29,7 +29,7 @@ public class FakeApiTest {
     /**
      * To test \&quot;client\&quot; model
      *
-     * 
+     * To test \&quot;client\&quot; model
      */
     @Test
     public void testClientModelTest() {
@@ -68,7 +68,7 @@ public class FakeApiTest {
     /**
      * To test enum parameters
      *
-     * 
+     * To test enum parameters
      */
     @Test
     public void testEnumParametersTest() {
@@ -78,7 +78,7 @@ public class FakeApiTest {
         String enumHeaderString = null;
         List<String> enumQueryStringArray = null;
         String enumQueryString = null;
-        BigDecimal enumQueryInteger = null;
+        Integer enumQueryInteger = null;
         Double enumQueryDouble = null;
         // api.testEnumParameters(enumFormStringArray, enumFormString, enumHeaderStringArray, enumHeaderString, enumQueryStringArray, enumQueryString, enumQueryInteger, enumQueryDouble);
